Fix: 0917 쿼리문 넣음

+++ demo/src/main/java/com/example/chusuk/demo/repository/PUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.chusuk.demo.entity.PUser;
 
-public interface PUserRepository extends JpaRepository<PUser, Long> {
+public interface PUserRepository extends JpaRepository<PUser, Integer> {
 
 }
